+++ .lintstagedrc.js
@@ -1,8 +1,8 @@
 module.exports = {
   '**/*.{ts,tsx,js,jsx,cjs,mjs}': [
-    'prettier --cache --write',
+    'rome format --write',
     'ts-standard --fix',
     'eslint --cache --fix'
   ],
-  '**/*.{json,md,yml,yaml}': ['prettier --cache --write']
+  '!(.vscode/**)**/*.{json,md,yml,yaml}': ['rome format --write']
 }