+++ tests/worker-files/thread/testMultiTasksWorker.js
@@ -8,9 +8,9 @@ const {
 
 module.exports = new ThreadWorker(
   {
-    jsonIntegerSerialization: data => jsonIntegerSerialization(data.n),
-    factorial: data => factorial(data.n),
-    fibonacci: data => fibonacci(data.n)
+    jsonIntegerSerialization: (data) => jsonIntegerSerialization(data.n),
+    factorial: (data) => factorial(data.n),
+    fibonacci: (data) => fibonacci(data.n)
   },
   {
     maxInactiveTime: 500,