+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.6.25] - 2023-08-13
+
 ### Added
 
 - HTTP server pool examples: fastify-cluster, fastify-hybrid.